chore: add frames to tabs and start debugger work
