+++ app/(drawer)/(tabs)/two.tsx
@@ -1,13 +1,10 @@
 import { Text, View } from 'react-native';
 
-import EditScreenInfo from '../../../components/edit-screen-info';
 
 export default function TabTwoScreen() {
   return (
     <View className={styles.container}>
-      <Text className={styles.title}>Tab Two</Text>
-      <View className={styles.separator} />
-      <EditScreenInfo path="app/(drawer)/(tabs)/two.tsx" />
+      <Text className={styles.title}>Frame Debugger</Text>
     </View>
   );
 }
